Add missing dependencies
