+++ README.md
@@ -9,6 +9,12 @@ HydrideSegmentation is a comprehensive toolkit for the segmentation and analysis
 - **Quantitative Analysis**: Computes hydride orientation, size distribution, and other morphological statistics.
 - **Publication-Quality Visualization**: Generates high-quality plots for hydride orientation maps, size distributions, and angle histograms.
 
+### Code Structure
+
+- `hydrideseg/gui_app.py` – main Tkinter application logic.
+- `hydrideseg/analysis.py` – utilities for hydride orientation analysis and plotting.
+- `GUI.py` – entry point with optional `--debug` mode.
+
 ## Example Workflow
 
 1. **Input Image**: Provide a microscopy image of zirconium hydrides.
@@ -44,12 +50,15 @@ Below is an example output generated by the toolkit, using the test image `test_
    ```bash
    python GUI.py
    ```
+   Running with `--debug` skips the GUI and saves a six-panel plot to
+   `test_data/results/3PB_SRT_data_generation_1817_OD_side1_8_combined.png`.
 
 ## Advanced Usage
 
 - **Segmentation Options**: Select between conventional and ML models in the GUI.
 - **Parameter Tuning**: Adjust CLAHE, thresholding, and morphological parameters for optimal segmentation.
 - **Debug Mode**: Use `--debug` for detailed logging and annotated plots.
+  The debug run creates a 2×3 figure combining segmentation and orientation analysis.
 
 ## Citation
