+++ requirements.txt
@@ -6,3 +6,6 @@ scikit-learn
 pillow
 albumentations
 tqdm
+
+scipy
+tkinterdnd2
